pyproject.toml
[project]
name = "lucid-docs"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Duarte",email = "p.roger.antunes43@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain[google-genai] (>=0.3.25,<0.4.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "fastapi[standard] (==0.115.9)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (==4.0.1)",
    "passlib (>=1.7.4,<2.0.0)",
]

[tool.poetry]
packages = [{include = "lucid_docs", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
